Simple program that uses an SQLite database to track job applications
First iteration of jobTracker. Features will be added in the future.
This will include the ability to import data into the table from a csv
and a seprate table for job seeking activities with the appropriate
interfaces. The icon is not mine, was found online and was added only as
practice and because the default icon is boring. I believe the icon is
originally from
http://blog.cloudera.com/blog/2009/04/hive-and-jobtracker-needed-logos/